Extract plan-analysis code to PerformanceMonitor.PlanAnalysis
Move ShowPlanParser, PlanAnalyzer, BenefitScorer, PlanLayoutEngine,
and PlanModels from Dashboard/{Services,Models} and Lite/{Services,Models}
into a new PerformanceMonitor.PlanAnalysis class library targeting
net10.0. Both apps now consume the shared plan group via
<ProjectReference>; ~5,100 LOC of byte-equivalent duplication
eliminated. Mirrors the pattern PR #1009 established for
PerformanceMonitor.Analysis. The two shared libraries are independent
(zero refs in either direction).

PlanAnalyzer.cs gains explicit `using System;`,
`using System.Collections.Generic;`, `using System.Linq;` at the top
because Lite's project-scoped GlobalUsings.cs isn't inherited by the
new project (which sets `<ImplicitUsings>disable</ImplicitUsings>`).
ScanImpact record adopts Dashboard's `sealed` modifier as the
reconciliation winner.

PlanIconMapper split to break an unanticipated incoming edge: the
plan classified PlanIconMapper as out-of-scope (WPF-bound, uses
BitmapImage), but ShowPlanParser calls PlanIconMapper.GetIconName at
parse time to populate PlanNode.IconName. The pure-data half
(IconMap dictionary + GetIconName lookup) is now `internal static
class IconNameMapper` inside the shared library; the per-app
PlanIconMapper.GetIcon (returns BitmapImage) stays per-app. The
per-app PlanIconMapper.GetIconName forwarder is removed — there
were no other callers and the user prefers no dead-code shims.

14 consumer files (7 Lite + 7 Dashboard) gain
`using PerformanceMonitor.PlanAnalysis;`. The two
ActualPlanExecutor-only consumers (Dashboard/Controls/
QueryPerformanceContent.Plans.cs and Lite/Controls/ServerTab.Plans.cs)
are intentionally not touched — ActualPlanExecutor stays per-app
this release because it calls ReproScriptBuilder (Class B, drifted
between Lite and Dashboard); both will be extracted in a follow-up
PR once ReproScriptBuilder is reconciled and a logging abstraction
is designed.

CHANGELOG updated. planalyzer-sync-checker agent retired (no copies
to sync).

Verified:
- dotnet build PerformanceMonitor.sln -c Debug: clean, 0 warnings.
- dotnet test Lite.Tests: 292 passed (matches origin/dev baseline).
- dotnet test Dashboard.Tests: 28 passed (matches origin/dev baseline).
- code-reviewer agent: no BLOCKING / MAJOR findings.
