+++ Dashboard/Analysis/SqlServerDrillDownCollector.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Data.SqlClient;
 using PerformanceMonitor.Analysis;
+using PerformanceMonitor.PlanAnalysis;
 using PerformanceMonitorDashboard.Helpers;
 using PerformanceMonitorDashboard.Mcp;
 using PerformanceMonitorDashboard.Models;
