+++ Dashboard/Controls/PlanViewerControl.Properties.cs
@@ -12,6 +12,7 @@ using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
+using PerformanceMonitor.PlanAnalysis;
 using PerformanceMonitorDashboard.Models;
 
 namespace PerformanceMonitorDashboard.Controls;
