+++ Dashboard/Dashboard.csproj
@@ -53,6 +53,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Installer.Core\Installer.Core.csproj" />
     <ProjectReference Include="..\PerformanceMonitor.Analysis\PerformanceMonitor.Analysis.csproj" />
+    <ProjectReference Include="..\PerformanceMonitor.PlanAnalysis\PerformanceMonitor.PlanAnalysis.csproj" />
   </ItemGroup>
 
   <ItemGroup>
