+++ Lite/Controls/PlanViewerControl.xaml.cs
@@ -14,6 +14,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
 using Microsoft.Win32;
+using PerformanceMonitor.PlanAnalysis;
 using PerformanceMonitorLite.Models;
 using PerformanceMonitorLite.Services;
