+++ PerformanceMonitor.PlanAnalysis/PerformanceMonitor.PlanAnalysis.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>disable</ImplicitUsings>
+    <LangVersion>latest</LangVersion>
+    <RootNamespace>PerformanceMonitor.PlanAnalysis</RootNamespace>
+    <AssemblyName>PerformanceMonitor.PlanAnalysis</AssemblyName>
+    <Company>Darling Data, LLC</Company>
+    <Copyright>Copyright © 2026 Darling Data, LLC</Copyright>
+    <EnableNETAnalyzers>true</EnableNETAnalyzers>
+    <AnalysisLevel>latest-recommended</AnalysisLevel>
+    <NoWarn>CA1849;CA2007;CA1508;CA1822;CA1805;CA1510;CA1816;CA1861;CA1845;CA2201;CA1848;CA1852;CA1305;CA1860;CA1707;CA1507;CA1806</NoWarn>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <InternalsVisibleTo Include="PerformanceMonitorLite" />
+    <InternalsVisibleTo Include="PerformanceMonitorDashboard" />
+    <InternalsVisibleTo Include="Lite.Tests" />
+    <InternalsVisibleTo Include="Dashboard.Tests" />
+  </ItemGroup>
+</Project>
